+++ src/Data/AppCategory.cs
@@ -1,0 +1,18 @@
+﻿namespace ChromaWrapper.Data
+{
+    /// <summary>
+    /// Specifies the category of a connected application.
+    /// </summary>
+    /// <seealso href="https://assets.razerzone.com/dev_portal/C%2B%2B/html/en/struct_chroma_s_d_k_1_1_a_p_p_i_n_f_o_t_y_p_e.html">ChromaSDK::APPINFOTYPE Struct Reference</seealso>
+    public enum AppCategory
+    {
+        /// <summary>None.</summary>
+        None = 0,
+
+        /// <summary>Utility application.</summary>
+        Utility = 1,
+
+        /// <summary>Game.</summary>
+        Game = 2,
+    }
+}