+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Update solution to .NET 6.0
+- Move enums `AppCategory` and `SupportedDevices` out of `ChromaAppInfo`
 
 ## [1.0.0](https://github.com/poveden/ChromaWrapper/releases/tag/v1.0.0) — 2021-10-09